Change Storyboarded to PackageStoryboarded
Also change the instantiate function to not take an identifier. In this
case where the package has View Controllers that conform to this
protocol, they should explicitly implement this function and provide the
appropriate identifier.

I think this will avoid any issues with using the "regular" Storyboarded
protocol in client projects.